Add CSS for Question component, move CSS to folder
