+++ src/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Provider } from 'react-redux'
 import logo from './logo.svg';
-import './App.css';
+import './css/App.css';
 import store from './store'
 import Game from './containers/Game'
