+++ src/css/Question.css
@@ -1,0 +1,30 @@
+.Question-img {
+    border: 5px solid #DD32E0;
+    box-sizing: border-box;
+    border-radius: 5px;
+    transform: rotate(-10deg);
+    max-height: 40vh;
+    max-width: 40vw;
+    margin: 10vh 0;
+}
+
+.Question-button {
+    background: linear-gradient(180deg, #FFFFFF 0%, rgba(255, 255, 255, 0.635417) 52.6%, rgba(255, 255, 255, 0) 100%), #DD32E0;
+    border: 3px solid #DD32E0;
+    box-sizing: border-box;
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+    border-radius: 5px;
+    font-size: 24px;
+    margin: 2vh 2vw;
+    padding: 10px;
+    transition: 0.3s;
+}
+
+.Question-button:hover {
+    transform: translateY(-5px)
+}
+
+.Question-button:active {
+    box-shadow: none;
+    transform: rotate(10deg)
+}
